add test database enviroment
