--- app/base.test.js
@@ -1,5 +1,0 @@
-describe('Initial Test', () => {
-  it('should test that 1 + 1 === 2', () => {
-    expect(1 + 1).toBe(2);
-  });
-});
