+++ app/database/models/index.js
@@ -1,25 +1,25 @@
 // const Pool = require('pg-pool');
 const { Sequelize } = require('sequelize');
 const { Pool } = require('pg');
-const dbConfig = require('../../config/db.config');
+const dbConfig = require('../../../config/db.config');
 const modelInitializer = require('./init-models');
 
 const sequelize = new Sequelize(
-  dbConfig.database,
-  dbConfig.user,
-  dbConfig.password,
-  {
-    host: dbConfig.host,
-    dialect: dbConfig.dialect,
-    operatorsAliases: false,
+	dbConfig.database,
+	dbConfig.user,
+	dbConfig.password,
+	{
+		host: dbConfig.host,
+		dialect: dbConfig.dialect,
+		operatorsAliases: false,
 
-    pool: {
-      max: dbConfig.pool.max,
-      min: dbConfig.pool.min,
-      acquire: dbConfig.pool.acquire,
-      idle: dbConfig.pool.idle,
-    },
-  },
+		pool: {
+			max: dbConfig.pool.max,
+			min: dbConfig.pool.min,
+			acquire: dbConfig.pool.acquire,
+			idle: dbConfig.pool.idle,
+		},
+	}
 );
 
 const models = modelInitializer.initModels(sequelize);
@@ -31,7 +31,7 @@ db.sequelize = sequelize;
 
 const modelNames = Object.keys(models);
 modelNames.forEach((modelKey) => {
-  db[modelKey] = models[modelKey];
+	db[modelKey] = models[modelKey];
 });
 
 module.exports = db;
