+++ app/database/models/related/related.model.js
@@ -1,0 +1,38 @@
+const Sequelize = require('sequelize');
+
+module.exports = (sequelize, DataTypes) =>
+	sequelize.define(
+		'related',
+		{
+			id: {
+				type: DataTypes.INTEGER,
+				allowNull: false,
+				primaryKey: true,
+			},
+			current_product_id: {
+				type: DataTypes.INTEGER,
+				allowNull: false,
+				references: {
+					model: 'product',
+					key: 'id',
+				},
+			},
+			related_product_id: {
+				type: DataTypes.INTEGER,
+				allowNull: true,
+			},
+		},
+		{
+			sequelize,
+			tableName: 'related',
+			schema: 'public',
+			timestamps: false,
+			indexes: [
+				{
+					name: 'related_pkey',
+					unique: true,
+					fields: [{ name: 'id' }],
+				},
+			],
+		}
+	);
