+++ app/database/seeders/20211001015107-product.js
@@ -1,0 +1,24 @@
+'use strict';
+
+module.exports = {
+	up: async (queryInterface, Sequelize) => {
+		/**
+		 * Add seed commands here.
+		 *
+		 * Example:
+		 * await queryInterface.bulkInsert('People', [{
+		 *   name: 'John Doe',
+		 *   isBetaMember: false
+		 * }], {});
+		 */
+	},
+
+	down: async (queryInterface, Sequelize) => {
+		/**
+		 * Add commands to revert seed here.
+		 *
+		 * Example:
+		 * await queryInterface.bulkDelete('People', null, {});
+		 */
+	},
+};
