+++ app/index.js
@@ -19,5 +19,6 @@ const server = app.listen(PORT, () => {
 });
 
 module.exports = {
+  app: server,
   close: () => server.close(),
 };
