--- app/models/style/style.model.test.js
@@ -1,30 +1,0 @@
-const {
-  sequelize,
-  dataTypes,
-  checkModelName,
-  checkUniqueIndex,
-  checkPropertyExists,
-  checkNonUniqueIndex,
-} = require('sequelize-test-helpers');
-
-const StyleModel = require('./style.model');
-
-const expectedFields = [
-  'id',
-  'name',
-  'original_price',
-  'sale_price',
-  'default?',
-  'style_id',
-];
-
-describe('Style Model', () => {
-  const Style = StyleModel(sequelize, dataTypes);
-  const style = new Style();
-
-  checkModelName(Style)('style');
-
-  describe('should contain expected properties', () => {
-    expectedFields.forEach(checkPropertyExists(style));
-  });
-});
