+++ package.json
@@ -5,9 +5,15 @@
 	"main": "index.js",
 	"scripts": {
 		"start": "nodemon app/index.js",
-		"test": "jest --no-cache --detectOpenHandles --runInBand --forceExit",
+		"test": "cross-env NODE_ENV=test jest --no-cache --detectOpenHandles --runInBand --forceExit --testTimeout=10000",
+		"test:watch": "cross-env NODE_ENV=test jest --no-cache --detectOpenHandles --runInBand --forceExit --testTimeout=10000 --onlyChanged --watch",
+		"pretest": "cross-env NODE_ENV=test npm run db:reset",
+		"db:create:test": "cross-env NODE_ENV=test npm run sequelize db:create",
+		"db:reset": "npm run sequelize db:drop && npm run sequelize db:create && npm run sequelize db:migrate && npm run sequelize db:seed:all",
 		"lint": "prettier --check .",
 		"format": "prettier --write .",
+		"sequelize": "sequelize --options-path=config/options.js",
+		"postinstall": "npm run db:create:test && npm run sequelize init:seeders && npm run sequelize init:migrations",
 		"csv:clean-file": "node csv_data/csvFileCleaner.js",
 		"csv:clean-all": "for i in csv_data/*.csv; do fullfilename=$i; filename=$(basename \"$fullfilename\"); fname=\"${filename%.*}\"; TARGET_FILE=\"$fname\" npm run csv:clean-file; done"
 	},
@@ -37,6 +43,7 @@
 	},
 	"devDependencies": {
 		"chai": "^4.3.4",
+		"cross-env": "^7.0.3",
 		"eslint": "^7.32.0",
 		"eslint-config-airbnb-base": "^14.2.1",
 		"eslint-plugin-import": "^2.24.2",
@@ -44,6 +51,7 @@
 		"jest": "^27.2.2",
 		"lint-staged": "^11.1.2",
 		"readline": "^1.3.0",
+		"sequelize-cli": "^6.2.0",
 		"sequelize-test-helpers": "^1.4.2",
 		"sinon": "^11.1.2",
 		"supertest": "^6.1.6"
@@ -66,6 +74,11 @@
 			"browser": true,
 			"es2021": true
 		},
+		"ignorePatterns": [
+			"models/",
+			"migrations/",
+			"seeders/"
+		],
 		"extends": [
 			"airbnb-base"
 		],
