+++ resources/scripts/sequelize-cli.sh
@@ -1,0 +1,2 @@
+npx sequelize-cli seed:generate --name projects --options-path=config/options.js
+npx sequelize-cli db:seed:all --options-path=config/options.js
